# [process-agent] reduce allocations in parseStatContent