--- a/releasenotes/notes/process-agent-improve-parse-stat-content-allocations-8c6fd48fb63649f1.yaml
+++ b/releasenotes/notes/process-agent-improve-parse-stat-content-allocations-8c6fd48fb63649f1.yaml
@@ -1,0 +1,3 @@
+---
+enhancements:
+  - Process-Agent: Improved parsing performance of the '/proc/pid/stat' file in (linux-only)