### What does this PR do?

This is a PR inspired by https://github.com/DataDog/datadog-agent/pull/20559 to reduce allocation when parsing the `/proc/$pid/stat` file. The following changes are introduced in the PR.
- Fill a pre-allocated buffer for each field. When the value matches a position of interest then perform non-allocating `[]byte` to numeral conversion. This reduces the allocation to zero from the original implmentation. 
- Break out of the loop after the last field of interest. This is about halfway through parsing of the file as there are about 52 fields in total according to: https://man7.org/linux/man-pages/man5/proc.5.html

### Benchmark Results

**Baseline**
```sh
BenchmarkParseStatContent-4       170738              6936 ns/op             544 B/op        108 allocs/op
PASS
ok      github.com/DataDog/datadog-agent/pkg/process/procutil   2.662s
```

**Use buffer and skip `[]byte` to string conversion**
```sh
BenchmarkParseStatContent-4       248408              4802 ns/op             224 B/op         12 allocs/op
PASS
ok      github.com/DataDog/datadog-agent/pkg/process/procutil   2.693s
```

**Pre-sizing buffer**
```sh
BenchmarkParseStatContent-4       243747              4768 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/DataDog/datadog-agent/pkg/process/procutil   2.896s
```

**Breaking out early from loop**
```sh
BenchmarkParseStatContent-4       485961              2352 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/DataDog/datadog-agent/pkg/process/procutil   2.869s
```

**Original vs Final**
```
name                old time/op    new time/op    delta
ParseStatContent-4    7.74µs ± 6%    2.30µs ± 4%   -70.22%  (p=0.000 n=9+9)

name                old alloc/op   new alloc/op   delta
ParseStatContent-4      544B ± 0%        0B       -100.00%  (p=0.000 n=10+10)

name                old allocs/op  new allocs/op  delta
ParseStatContent-4       108 ± 0%         0       -100.00%  (p=0.000 n=10+10)
```

### Motivation

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [x] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [x] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [x] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [x] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature.
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.